Fleshed out Motor Control module. Still not working properly.
Debugging shows that either MTR is only posedged once, or MESSAGE does
not change after initial transmission (i.e. only works once).
Investigate this.

Additionally, nothing seems to be showing up on the Arduino end of the
SPI. Likely an issue with my connection to the SPIM module. It could
possibly be an issue with the Arduino code. This part can be dismissed
by using another Arduino as an SPI Master. Investigate this.